+++ src/iGameExtern.h
@@ -85,7 +85,7 @@ typedef struct settings
 	int no_guigfx;
 	int start_with_favorites;
 	int useIgameDataTitle;
-	uint8_t lastScanSetup;  // It keeps info of the settings on the last scan
+	int lastScanSetup;  // It keeps info of the settings on the last scan
 							// that influence the item data
 							// It is a bitfield with the following structure
 							//  ------------ no_smart_spaces
